Allow exclusion of references for git LFS fetch
Prefixing a ref name with a caret ^ allows exclusion of refs.
This is very useful when using the lfs prune --force command as
git lfs fetch would normally download all files of the current checkout.

To download all files that are new on the remote call:
git lfs fetch origin origin/main ^main