Add a safe option to LFS prune
LFS prune offers --verify-remote, a command that exits whenever a reachable object is not on the remote. Hence, it still prunes non-reachable objects, so objects in the index or files referenced only by orphan commits.
Deleting LFS objects in the cache is a destructive operation that is undoable.
This commit adds a new option --safe that guarantees to only delete files that are available on the remote, so there is a guarantee that data is never lost.